+++ Commons/UICore/src/main/java/com/mfino/uicore/fix/processor/impl/GetAvialableBalanceProcessorImpl.java
@@ -48,7 +48,7 @@ public class GetAvialableBalanceProcessorImpl extends FIXMessageHandler implemen
 	    	balanceInquiry.setSourceMDN(subscriberService.normalizeMDN(realMsg.getSourceMDN()));
 	    	balanceInquiry.setPin(realMsg.getPin());
 	    	balanceInquiry.setServletPath(CmFinoFIX.ServletPath_BankAccount);
-	    	balanceInquiry.setSourceApplication(((Long)cc.getChannelsourceapplication()).intValue());
+	    	balanceInquiry.setSourceApplication((cc.getChannelsourceapplication()).intValue());
 	    	balanceInquiry.setChannelCode(cc.getChannelcode());
 	    	balanceInquiry.setPocketID(realMsg.getSourcePocketID());
