+++ Core/test/com/mfino/dao/CommodityTransferDAOTest.java
@@ -214,12 +214,12 @@ public class CommodityTransferDAOTest {
         pcmt.setStarttime(new Timestamp());
         pcmt.setSourcemdn(mdn1.getMdn());
         pcmt.setBankauthorizationcode("sunny");
-        pcmt.setBankcode(Long.MAX_VALUE);
+        pcmt.setBankcode(Integer.MAX_VALUE);
         pcmt.setBankerrortext("sunny");
         pcmt.setBankrejectreason("sunnys");
-        pcmt.setBankresponsecode(Long.MAX_VALUE);
+        pcmt.setBankresponsecode(Integer.MAX_VALUE);
         pcmt.setBankresponsetime(new Timestamp());
-        pcmt.setBillingtype(Long.MAX_VALUE);
+        pcmt.setBillingtype(Integer.MAX_VALUE);
         pcmt.setBuckettype("sunnys");
         pcmt.setCommodity(Integer.MAX_VALUE);
         pcmt.setCreatetime(new Timestamp());
@@ -227,7 +227,7 @@ public class CommodityTransferDAOTest {
         pcmt.setSourcecardpan("bunny");
         pcmt.setDestmdnid(mdn2.getId());
         pcmt.setDestpocketbalance(new BigDecimal(Long.MIN_VALUE)+"");
-        pcmt.setDestpockettype(Long.MAX_VALUE);
+        pcmt.setDestpockettype(Integer.MAX_VALUE);
         pcmt.setDestpocketid(new BigDecimal(Long.MIN_VALUE));
         pcmt.setDestsubscriberid(new BigDecimal(Long.MIN_VALUE));
         pcmt.setDestsubscribername("sunny");
@@ -235,10 +235,10 @@ public class CommodityTransferDAOTest {
         pcmt.setLastupdatetime(new Timestamp());
         pcmt.setMsgtype(Integer.MAX_VALUE);
         pcmt.setOperatorauthorizationcode("ssunny");
-        pcmt.setOperatorcode(Long.MAX_VALUE);
+        pcmt.setOperatorcode(Integer.MAX_VALUE);
         pcmt.setOperatorerrortext("Error");
         pcmt.setOperatorrejectreason("sunnys");
-        pcmt.setOperatorresponsecode(Long.MAX_VALUE);
+        pcmt.setOperatorresponsecode(Integer.MAX_VALUE);
         pcmt.setOperatorresponsetime(new Timestamp());
         pcmt.setSourceapplication(Integer.MAX_VALUE);
         pcmt.setSourcecardpan("sda");
@@ -246,7 +246,7 @@ public class CommodityTransferDAOTest {
         pcmt.setSourcepockettype(Integer.MAX_VALUE);
         pcmt.setSourcereferenceid("dssd");
         pcmt.setSourcesubscribername("sunnys");
-        pcmt.setTransferfailurereason(Long.MAX_VALUE);
+        pcmt.setTransferfailurereason(Integer.MAX_VALUE);
         pcmt.setTransferstatus(Integer.MAX_VALUE);
         pcmt.setIso8583Primaryaccountnumber("12345");
 
@@ -261,7 +261,7 @@ public class CommodityTransferDAOTest {
         pcmt.setPocket(poc);
         pcmt.setTransactionLog(tlog);
         pcmt.setLastreversaltime(new Timestamp());
-        pcmt.setReversalcount(1L);
+        pcmt.setReversalcount(1);
 
 
         dao.save(pcmt);
@@ -273,12 +273,12 @@ public class CommodityTransferDAOTest {
         pcmt1.setStarttime(new Timestamp());
         pcmt1.setSourcemdn(mdn2.getMdn());
         pcmt1.setBankauthorizationcode("sunny");
-        pcmt1.setBankcode(Long.MAX_VALUE);
+        pcmt1.setBankcode(Integer.MAX_VALUE);
         pcmt1.setBankerrortext("sunny");
         pcmt1.setBankrejectreason("sunnys");
-        pcmt1.setBankresponsecode(Long.MAX_VALUE);
+        pcmt1.setBankresponsecode(Integer.MAX_VALUE);
         pcmt1.setBankresponsetime(new Timestamp());
-        pcmt1.setBillingtype(Long.MAX_VALUE);
+        pcmt1.setBillingtype(Integer.MAX_VALUE);
         pcmt1.setBuckettype("sunnys");
         pcmt1.setCommodity(Integer.MAX_VALUE);
         pcmt1.setCreatetime(new Timestamp());
@@ -286,7 +286,7 @@ public class CommodityTransferDAOTest {
         pcmt1.setSourcecardpan("bunny");
         pcmt1.setDestmdnid(mdn1.getId());
         pcmt1.setDestpocketbalance(new BigDecimal(Long.MIN_VALUE)+"");
-        pcmt1.setDestpockettype(Long.MAX_VALUE);
+        pcmt1.setDestpockettype(Integer.MAX_VALUE);
         pcmt1.setDestpocketid(new BigDecimal(Long.MIN_VALUE));
         pcmt1.setDestsubscriberid(new BigDecimal(Long.MIN_VALUE));
         pcmt1.setDestsubscribername("sunny");
@@ -294,10 +294,10 @@ public class CommodityTransferDAOTest {
         pcmt1.setLastupdatetime(new Timestamp());
         pcmt1.setMsgtype(Integer.MAX_VALUE);
         pcmt1.setOperatorauthorizationcode("ssunny");
-        pcmt1.setOperatorcode(Long.MAX_VALUE);
+        pcmt1.setOperatorcode(Integer.MAX_VALUE);
         pcmt1.setOperatorerrortext("Error");
         pcmt1.setOperatorrejectreason("sunnys");
-        pcmt1.setOperatorresponsecode(Long.MAX_VALUE);
+        pcmt1.setOperatorresponsecode(Integer.MAX_VALUE);
         pcmt1.setOperatorresponsetime(new Timestamp());
         pcmt1.setSourceapplication(Integer.MAX_VALUE);
         pcmt1.setSourcecardpan("sdahhkjhjk");
@@ -305,7 +305,7 @@ public class CommodityTransferDAOTest {
         pcmt1.setSourcepockettype(Integer.MAX_VALUE);
         pcmt1.setSourcereferenceid("dssd");
         pcmt1.setSourcesubscribername("sunnys");
-        pcmt1.setTransferfailurereason(Long.MAX_VALUE);
+        pcmt1.setTransferfailurereason(Integer.MAX_VALUE);
         pcmt1.setTransferstatus(Integer.MAX_VALUE);
         pcmt1.setIso8583Primaryaccountnumber("12345");
 
@@ -319,7 +319,7 @@ public class CommodityTransferDAOTest {
         pcmt1.setPocket(poc);
         pcmt1.setTransactionLog(tlog);
         pcmt1.setLastreversaltime(new Timestamp());
-        pcmt1.setReversalcount(1L);
+        pcmt1.setReversalcount(1);
 
         dao.save(pcmt1);
